test(fuzz): add MAX_READ_BUF_LEN
No bounds on buffer size. Can you hear the allocator's cries?

+++ fuzz/src/method.rs
@@ -4,7 +4,7 @@ use super::SeekFrom;
 
 #[derive(arbitrary::Arbitrary, Debug)]
 pub enum Method {
-    Read { buf_len: usize },
+    Read { buf_len: u32 },
     Write { buf: Vec<u8> },
     Seek { seek_from: SeekFrom },
 }
